Minor editions
Apostrophes should not be used in formal writing (no problem about conversations), numbers neither, and I don't know what to make of that, Mach is the speed of sound, which can't travel in the void. So technically you can say a spaceship is going Mach 23, but in practice you would use just kms/h.